Use yandex search without map controller
