+++ lib/yandex_mapkit.dart
@@ -8,6 +8,7 @@ export 'src/polyline.dart';
 export 'src/suggest_item.dart';
 export 'src/yandex_map.dart';
 export 'src/yandex_map_controller.dart';
+export 'src/yandex_search.dart';
 
 class YandexMapkit {
 }
